# src/main/resources/api/api.yaml
openapi: 3.0.3
info:
  title: Lawyer
  description: Lawyer
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: localhost
paths:
  /addresses:
    get:
      tags:
        - Addresses
      description: returns list of addresses
      responses:
        200:
          description: list of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags:
        - Addresses
      description: post address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        200:
          description: ok
    put:
      tags:
        - Addresses
      description: put address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        200:
          description: ok

  /addresses/{id}:
    parameters:
      - name: id
        in: path
        description: id of address
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Addresses
      description: returns single address
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Address'
    delete:
      tags:
        - Addresses
      description: deletes single address
      responses:
        200:
          description: ok

  /contacts:
    get:
      tags:
        - Contacts
      description: returns list of contacts
      responses:
        200:
          description: list of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
    post:
      tags:
        - Contacts
      description: post contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Contact'
      responses:
        200:
          description: ok
    put:
      tags:
        - Contacts
      description: put contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Contact'
      responses:
        200:
          description: ok

  /contacts/{id}:
    parameters:
      - name: id
        in: path
        description: id of contact
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Contacts
      description: returns single contact
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Contact'
    delete:
      tags:
        - Contacts
      description: deletes single contact
      responses:
        200:
          description: ok

  /courts:
    get:
      tags:
        - Courts
      description: returns list of courts
      responses:
        200:
          description: list of courts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Court'
    post:
      tags:
        - Courts
      description: post court
      requestBody:
        required: true
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Court'
      responses:
        200:
          description: ok
    put:
      tags:
        - Courts
      description: put court
      requestBody:
        required: true
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Court'
      responses:
        200:
          description: ok

  /courts/{id}:
    parameters:
      - name: id
        in: path
        description: id of court
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Courts
      description: returns single court
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Court'
    delete:
      tags:
        - Courts
      description: deletes single court
      responses:
        200:
          description: ok

  /lawsuits:
    get:
      tags:
        - Lawsuits
      description: returns list of lawsuits
      responses:
        200:
          description: list of lawsuits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lawsuit'
    post:
      tags:
        - Lawsuits
      description: post lawsuit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lawsuit'
      responses:
        200:
          description: ok
    put:
      tags:
        - Lawsuits
      description: put lawsuit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lawsuit'
      responses:
        200:
          description: ok

  /lawsuits/{id}:
    parameters:
      - name: id
        in: path
        description: id of lawsuit
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Lawsuits
      description: returns single lawsuit
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Lawsuit'
    delete:
      tags:
        - Lawsuits
      description: deletes single lawsuit
      responses:
        200:
          description: ok

  /notes:
    get:
      tags:
        - Notes
      description: returns list of notes
      responses:
        200:
          description: list of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
    post:
      tags:
        - Notes
      description: post note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        200:
          description: ok
    put:
      tags:
        - Notes
      description: put note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        200:
          description: ok

  /notes/{id}:
    parameters:
      - name: id
        in: path
        description: id of note
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Notes
      description: returns single note
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Note'
    delete:
      tags:
        - Notes
      description: deletes single note
      responses:
        200:
          description: ok

  /tags:
    get:
      tags:
        - Tags
      description: returns list of tags
      responses:
        200:
          description: list of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      tags:
        - Tags
      description: post tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        200:
          description: ok
    put:
      tags:
        - Tags
      description: put tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        200:
          description: ok

  /tags/{id}:
    parameters:
      - name: id
        in: path
        description: id of tag
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Tags
      description: returns single tag
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Tag'
    delete:
      tags:
        - Tags
      description: deletes single tag
      responses:
        200:
          description: ok

  /tasks:
    get:
      tags:
        - Tasks
      description: returns list of tasks
      responses:
        200:
          description: list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      tags:
        - Tasks
      description: post task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        200:
          description: ok
    put:
      tags:
        - Tasks
      description: put task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        200:
          description: ok

  /tasks/{id}:
    parameters:
      - name: id
        in: path
        description: id of task
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Tasks
      description: returns single task
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Task'
    delete:
      tags:
        - Tasks
      description: deletes single task
      responses:
        200:
          description: ok

components:
  schemas:
    Address:
      properties:
        id:
          type: string
          format: uuid
        street:
          type: string
        zipCode:
          type: string
        city:
          type: string
        country:
          type: string
      required:
        - id
        - street
        - city

    Contact:
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        altEmail:
          type: string
          format: email
        phone:
          type: string
        altPhone:
          type: string
        companyName:
          type: string
        pesel:
          type: string
        nip:
          type: string
        regon:
          type: string
        krs:
          type: string
        dateCreated:
          type: string
          format: date-time
        address:
          $ref: '#/components/schemas/Address'
        correspondenceAddress:
          $ref: '#/components/schemas/Address'

    Court:
      properties:
        courtId:
          type: string
          format: uuid
        name:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        description:
          type: string
        phone:
          type: string

    Lawsuit:
      properties:
        lawsuitId:
          type: string
          format: uuid
        name:
          type: string
        caseSide:
          type: string
        inpuntDate:
          type: string
          format: date
        deadline:
          type: string
          format: date
        signature:
          type: string
        claimAmount:
          type: number
        sdditionalInfo:
          type: string
        contactList:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        taskList:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        plaintiff:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        defendant:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        eventSet:
          type: array
          items:
            $ref: '#/components/schemas/Event'

    Note:
      properties:
        noteId:
          type: string
          format: uuid
        title:
          type: string
        text:
          type: string

    Tag:
      properties:
        tagId:
          type: string
          format: uuid
        name:
          type: string

    Task:
      properties:
        taskId:
          type: string
          format: uuid
        priority:
          type: boolean
        deadline:
          type: string
          format: date-time
        description:
          type: string
        contactList:
          type: array
          items:
            $ref: '#/components/schemas/Address'

    Event:
      properties:
        eventId:
          type: string
          format: uuid
        title:
          type: string
        dateTime:
          type: string
          format: date-time
        description:
          type: string

    Payment:
      properties:
        paymentId:
          type: string
          format: uuid
        paymentValue:
          type: number
        paymentDate:
          type: string
          format: date
        paid:
          type: boolean
        paidDate:
          type: string
          format: date
        comment:
          type: string
        us:
          type: boolean
        incoming:
          type: boolean

    Poa:
      properties:
        poaId:
          type: string
          format: uuid
        poaType:
          type: string
        payment:
          type: string
        kpc:
          type: boolean
        termination:
          type: boolean
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        terminationNotificationDuty:
          type: boolean
        terminationNotificationDutyCompleted:
          type: boolean